Fix: API 버그 수정 및 코드 품질 개선
  - SignalCycleRepository: numeric → double precision 캐스팅으로 타입 변환 오류 수정
  - SignalCycleRepository: JdbcTemplate deprecated API를 varargs 방식으로 변경
  - GoogleAuthService, KakaoAuthService: getBody() null 반환 시 InvalidTokenException throw
  - RouteService: resolveSignalState() null guard 추가
  - RouteService: AI 경로 인덱스 범위 초과 시 명시적 fallback 처리
  - RouteResponse: nodeId 타입 long → int (IntersectionInfo.itstId()와 일치)
  - BusStopController: radiusM 음수 방지를 위한 @Validated, @DecimalMin(0) 추가
  - JwtProviderTest: @PostConstruct init() 수동 호출 추가
  - RouteServiceTest: fetchAll 시그니처 및 누락 Mock 반영

+++ backend/src/main/java/kr/io/pacer/core/dto/response/RouteResponse.java
@@ -64,7 +64,7 @@ public class RouteResponse {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class SignalCheckpoint {
-        private long            nodeId;
+        private int             nodeId;
         private double          lat;
         private double          lng;
         private int             etaFromStartSeconds;
